+++ CHANGELOG.md
@@ -5,3 +5,9 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## [Unreleased]
 - Initial release
+
+### [1.0.0]
+- Initial release of Comment Label extension.
+
+### [1.1.0]
+- Added rudimentary support for more common languages.